works but fails 3 tests

+++ blogicum/blog/views.py
@@ -1,17 +1,16 @@
-from django.contrib.auth import login
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
+from django.core.exceptions import PermissionDenied
 from django.core.mail import send_mail
 from django.core.paginator import Paginator
 from django.db.models import Count
-from django.views.generic import CreateView, UpdateView, ListView, DetailView
 from django.shortcuts import get_object_or_404, render, redirect
-from django.urls import reverse_lazy
+from django.urls import reverse
 from django.utils import timezone
 
 from .constants import NUMBER_POSTS
-from .form import PostForm, RegistrationForm, CommentForm, StaticPageForm
-from .models import Category, Post, Comment, StaticPage
+from .form import PostForm, RegistrationForm, CommentForm
+from .models import Category, Post, Comment
 
 
 def get_filter_posts(author=None, location=None):
@@ -49,7 +48,7 @@ def post_detail(request, post_id):
         get_filter_posts(),
         id=post_id,
     )
-    comments = post.comment.all()
+    comments = post.comment.all().order_by('created_at')
     form = CommentForm()
     return render(request, 'blog/detail.html', {
         'post': post,
@@ -68,10 +67,14 @@ def category_posts(request, category_slug):
     author = request.GET.get('author')
     location = request.GET.get('location')
 
-    page_obj = get_filter_posts(
+    posts = get_filter_posts(
         author=author,
         location=location
-    ).filter(category=category)
+    ).filter(category=category).order_by('-pub_date')
+
+    paginator = Paginator(posts, NUMBER_POSTS)
+    page_number = request.GET.get('page')
+    page_obj = paginator.get_page(page_number)
 
     context = {
         'category': category_slug,
@@ -80,7 +83,7 @@ def category_posts(request, category_slug):
     return render(request, 'blog/category.html', context)
 
 
-@login_required
+@login_required(login_url='/login/')
 def create_post(request):
     form = PostForm()
     if request.method == 'POST':
@@ -94,31 +97,37 @@ def create_post(request):
     return render(request, 'blog/create.html', {'form': form})
 
 
-@login_required
+@login_required(login_url='/login/')
 def edit_post(request, post_id):
     post = get_object_or_404(Post, id=post_id)
 
     if post.author != request.user:
-        return redirect('blog:post_detail', post_id=post.id)
+        return redirect(reverse(
+            'blog:post_detail',
+            kwargs={'post_id': post.id})
+        )
 
     if request.method == 'POST':
         form = PostForm(request.POST, request.FILES, instance=post)
         if form.is_valid():
             form.save()
-            return redirect('blog:post_detail', post_id=post.id)
+            return redirect(reverse(
+                'blog:post_detail',
+                kwargs={'post_id': post.id})
+            )
     else:
         form = PostForm(instance=post)
 
-    return render(request, 'blog/create.html', {'form': form})
+    return render(request, 'blog/create.html', {'form': form, 'post': post})
 
 
-@login_required
+@login_required(login_url='/login/')
 def delete_post(request, post_id):
     post = get_object_or_404(Post, id=post_id)
 
     if post.author == request.user:
         post.delete()
-        return redirect('blog:profile', username=request.user.username)
+        return redirect('blog:index', username=request.user.username)
     else:
         return redirect('blog:post_detail', post_id=post.id)
 
@@ -128,7 +137,7 @@ def profile(request, username):
     posts = Post.objects.filter(author=user_profile).annotate(
         comment_count=Count('comment')
     ).order_by('-pub_date')
-    paginator = Paginator(posts, 10)
+    paginator = Paginator(posts, NUMBER_POSTS)
     page_number = request.GET.get('page')
     page_obj = paginator.get_page(page_number)
     context = {
@@ -138,7 +147,7 @@ def profile(request, username):
     return render(request, 'blog/profile.html', context)
 
 
-@login_required
+@login_required(login_url='/login/')
 def edit_profile(request):
     user = request.user
 
@@ -157,23 +166,7 @@ def edit_profile(request):
     )
 
 
-@login_required
-def register_view(request):
-    if request.method == 'POST':
-        form = RegistrationForm(request.POST)
-        if form.is_valid():
-            user = form.save()
-            login(request, user)
-            return redirect('blog:profile', username=user.username)
-    else:
-        form = RegistrationForm()
-    return render(
-        request,
-        'registration/registration_form.html',
-        {'form': form}
-    )
-
-
+@login_required(login_url='/login/')
 def add_comment(request, post_id):
     post = get_object_or_404(Post, id=post_id)
 
@@ -191,35 +184,45 @@ def add_comment(request, post_id):
     return render(request, 'blog/comment.html', {'post': post, 'form': form})
 
 
-def edit_comment(request, comment_id):
+@login_required(login_url='/login/')
+def edit_comment(request, post_id, comment_id):
     comment = get_object_or_404(Comment, id=comment_id)
 
     if comment.author != request.user:
-        return redirect('blog:post_detail', post_id=comment.post.id)
+        raise PermissionDenied(
+            "У вас нет прав для редактирования данного комментария."
+        )
+    form = CommentForm(instance=comment)
 
     if request.method == 'POST':
         form = CommentForm(request.POST, instance=comment)
         if form.is_valid():
             form.save()
-            return redirect('blog:post_detail', post_id=comment.post.id)
-    else:
-        form = CommentForm(instance=comment)
+            return redirect('blog:post_detail', post_id=comment.post.pk)
 
     return render(
-        request, 'blog/edit_comment.html',
+        request, 'blog/comment.html',
         {'form': form, 'comment': comment}
     )
 
 
-def delete_comment(request, comment_id):
+@login_required(login_url='/login/')
+def delete_comment(request, post_id, comment_id):
     comment = get_object_or_404(Comment, id=comment_id)
+    # form = CommentForm(request.POST, instance=comment)
+    # context = {'form': CommentForm()}
 
     if comment.author == request.user:
-        post_id = comment.post.id
-        comment.delete()
-        return redirect('blog:post_detail', post_id=post_id)
+        if request.method == 'GET':
+            comment.delete()
+            return redirect('blog:index')
     else:
-        return redirect('blog:post_detail', post_id=comment.post.id)
+        raise PermissionDenied(
+            "У вас нет прав для удаления данного комментария."
+        )
+
+    return redirect('blog:post_detail', post_id=post_id)
+    # return render(request, 'blog/comment.html', context)
 
 
 def confirm_create_post(email):
@@ -230,35 +233,3 @@ def confirm_create_post(email):
         recipient_list=[email],
         fail_silently=True,
     )
-
-
-class StaticPageListView(ListView):
-    """Шаблон для списка страниц"""
-
-    model = StaticPage
-    template_name = 'static_pages/static_page_list.html'
-
-
-class StaticPageDetailView(DetailView):
-    """Шаблон для отображения страницы"""
-
-    model = StaticPage
-    template_name = 'static_pages/static_page_detail.html'
-
-
-class StaticPageCreateView(CreateView):
-    """Перенаправление после успешного создания"""
-
-    model = StaticPage
-    form_class = StaticPageForm
-    template_name = 'static_pages/static_page_form.html'
-    success_url = reverse_lazy('static_page_list')
-
-
-class StaticPageUpdateView(UpdateView):
-    """Перенаправление после успешного обновления"""
-
-    model = StaticPage
-    form_class = StaticPageForm
-    template_name = 'static_pages/static_page_form.html'
-    success_url = reverse_lazy('static_page_list')
